7039 Move NavigateFromMainScreen to InstanceConfigurationScreenMockTest
